reword! Spinning up the hamster...
DISCLAIMER: this commit message has been retrieved from whatthecommit.com.

+++ src/rest_model.rs
@@ -509,9 +509,7 @@ pub enum OrderSide {
 
 /// By default, buy
 impl Default for OrderSide {
-    fn default() -> Self {
-        Self::Buy
-    }
+    fn default() -> Self { Self::Buy }
 }
 
 /// The allowed values are:
@@ -526,9 +524,7 @@ pub enum CancelReplaceMode {
 
 /// By default, STOP_ON_FAILURE
 impl Default for CancelReplaceMode {
-    fn default() -> Self {
-        Self::StopOnFailure
-    }
+    fn default() -> Self { Self::StopOnFailure }
 }
 
 /// Order types, the following restrictions apply
@@ -555,9 +551,7 @@ pub enum OrderType {
 
 /// By default, use market orders
 impl Default for OrderType {
-    fn default() -> Self {
-        Self::Market
-    }
+    fn default() -> Self { Self::Market }
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
@@ -567,7 +561,7 @@ pub struct MarginOrder {
     pub side: OrderSide,
     #[serde(rename(serialize = "type", deserialize = "type"))]
     pub order_type: OrderType,
-    pub quantity: Option<f64>,
+    pub quantity: Option<f32>,
     pub quote_order_qty: Option<f64>,
     pub price: Option<f64>,
     /// Used with `OrderType::StopLoss`, `OrderType::StopLossLimit`, `OrderType::TakeProfit` and `OrderType::TakeProfitLimit`
